+++ bin/devkit.php
@@ -3,6 +3,7 @@
 // @link https://www.tomasvotruba.cz/blog/2018/08/02/5-gotchas-of-the-bin-file-in-php-cli-applications/
 
 use Bartlett\PHPToolbox\Console\ApplicationInterface;
+use Bartlett\PHPToolbox\Console\Command\About;
 use Bartlett\PHPToolbox\Console\Command\UpdateExtensions;
 use Bartlett\PHPToolbox\Console\Command\UpdateTools;
 use Bartlett\PHPToolbox\DependencyInjection\ContainerFactory;
@@ -11,6 +12,7 @@ require_once dirname(__DIR__) . '/config/bootstrap.php';
 
 $container = (new ContainerFactory())->create(basename(__FILE__, '.php'));
 $application = $container->get(ApplicationInterface::class);
+$application->add(new About());
 $application->add(new UpdateExtensions());
 $application->add(new UpdateTools());
 exit($application->run());