+++ src/Console/Application.php
@@ -7,8 +7,13 @@
  */
 namespace Bartlett\PHPToolbox\Console;
 
+use Composer\InstalledVersions;
+
 use Symfony\Component\Console\Application as SymfonyApplication;
 use Symfony\Component\DependencyInjection\ContainerInterface;
+
+use function sprintf;
+use function substr;
 
 /**
  * @since Release 1.0.0alpha1
@@ -23,7 +28,10 @@ final class Application extends SymfonyApplication implements ApplicationInterfa
      */
     public function __construct()
     {
-        parent::__construct(self::NAME, self::VERSION);
+        parent::__construct(
+            self::NAME,
+            $this->getInstalledVersion(false)
+        );
     }
 
     /**
@@ -33,5 +41,37 @@ final class Application extends SymfonyApplication implements ApplicationInterfa
     public function setContainer(ContainerInterface $container = null)
     {
         $this->container = $container;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getHelp(): string
+    {
+        return sprintf(
+            '<info>%s</info> version <comment>%s</comment>',
+            $this->getName(),
+            $this->getVersion()
+        );
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getLongVersion(): string
+    {
+        return $this->getInstalledVersion();
+    }
+
+    public function getInstalledVersion(bool $withRef = true): string
+    {
+        $packageName = 'bartlett/docker-php-toolbox';
+
+        $version = InstalledVersions::getPrettyVersion($packageName);
+        if (!$withRef) {
+            return $version;
+        }
+        $commitHash = InstalledVersions::getReference($packageName);
+        return sprintf('%s@%s', $version, substr($commitHash, 0, 7));
     }
 }